don't sleep bot empty walls

+++ core/src/com/mygdx/game/Map.java
@@ -21,8 +21,8 @@ public class Map {
         this.obstacles = new int[SIZE_X][SIZE_Y];
         for (int i = 0; i < SIZE_X; i++) {
             for (int j = 0; j < SIZE_Y; j++) {
-                int cx = (int)(i / 4);
-                int cy = (int)(j / 4);
+                int cx = i / 4;
+                int cy = j / 4;
                 if (cx % 2 == 0 && cy % 2 == 0) {
                     obstacles[i][j] = 5;
                 }
